better logic for /var/www/html/index.html

+++ deploy/custom_deploy_source.bash
@@ -4,7 +4,7 @@ post_apache_deploy() {
 }
 mk_index_html() {
   local tmpf=$(mktemp)
-  [[ -e /var/www/html/index.html ]] || cat > $tmpf <<'EOF'
+  cat > $tmpf <<'EOF'
 <html>
   <head>
     <meta http-equiv="refresh" content="0; url=app.html" />
@@ -18,7 +18,7 @@ mk_index_html() {
 </html>
 EOF
   sudo chown $apacheu $tmpf
-  sudo -u $apacheu mv $tmpf /var/www/html/index.html
+  [[ -e /var/www/html/index.html ]] || sudo -u $apacheu mv $tmpf /var/www/html/index.html
 }
 post_db_deploy() {
   mysql my_db1 -B -v -e "INSERT INTO users SET id='$DBU', nick='$DBU' ON DUPLICATE KEY UPDATE nick='$DBU'"
